fix: count results done
