+++ frontend/public/off.html
@@ -179,7 +179,7 @@
                     <div class="small-12 columns filterProducts">
                       <div class="filterProducts__buttons">
                         <div>
-                          <results-counter-display></results-counter-display>
+                          <searchalicious-count search-name="off"></searchalicious-count>
                         </div>
                         <div>
                           <button
