+++ frontend/src/search-a-licious.ts
@@ -3,4 +3,4 @@ export {SearchaliciousButton} from './search-button';
 export {SearchaliciousPages} from './search-pages';
 export {SearchaliciousFacets} from './search-facets';
 export {SearchaliciousResults} from './search-results';
-export {ResultsCounterDisplay} from './results-counter-display';
+export {ResultsCounterDisplay} from './search-count';
